# app/reports.py
from flask import Blueprint, Response, current_app, flash, jsonify, redirect, render_template, request, session, stream_with_context, url_for
from sqlalchemy import func, select

from app import db, login_required
from app.models import Area, Attachment, FuelLoad, FuelLoadGenerator, Machinery, Report, ReportChecklist, Section, User
from app.services.field_reports import (
    BusinessRuleError, CHECKLIST, IMAGE_EXTENSIONS, SERVICE_TYPES,
    create_fuel_load as create_fuel_load_service,
    create_maintenance_report,
)
from app.services.storage import StorageError, get_file_storage, resolve_fuel_key, resolve_report_key

reports_bp = Blueprint("reports", __name__)


def _can_access_record(technician_id):
    return session.get("role") == "Administrador" or technician_id == session.get("user_id")


def _file_response(storage, key, filename, mime_type, *, attachment=True):
    if not storage.exists(key):
        return "Archivo no encontrado.", 404
    response = Response(stream_with_context(storage.iter_bytes(key)), mimetype=mime_type)
    response.headers.set("Content-Disposition", "attachment" if attachment else "inline", filename=filename)
    response.headers["X-Content-Type-Options"] = "nosniff"
    return response
@reports_bp.get("/")
@login_required
def dashboard():
    query = _filtered_reports().order_by(Report.created_at.desc())
    reports = db.session.execute(query).scalars().all()
    # SQL Server usa "= 1" para BIT; "IS 1" no es sintaxis válida.
    technicians = db.session.execute(select(User).where(User.is_active == True).order_by(User.full_name)).scalars().all()  # noqa: E712
    return render_template("dashboard.html", reports=reports, technicians=technicians, areas=db.session.execute(select(Area).order_by(Area.name)).scalars().all(), sections=db.session.execute(select(Section).order_by(Section.name)).scalars().all(), machines=db.session.execute(select(Machinery).order_by(Machinery.name)).scalars().all())


def _filtered_reports():
    query = select(Report)
    client = request.args.get("client", "").strip()
    for column, value in ((Report.technician_id, request.args.get("technician_id", type=int)), (Report.area_id, request.args.get("area_id", type=int)), (Report.section_id, request.args.get("section_id", type=int)), (Report.machinery_id, request.args.get("machinery_id", type=int))):
        if value: query = query.where(column == value)
    if client: query = query.where(Report.client.ilike(f"%{client}%"))
    if request.args.get("date_from"): query = query.where(Report.created_at >= request.args["date_from"])
    if request.args.get("date_to"): query = query.where(Report.created_at < f"{request.args['date_to']} 23:59:59.9999999")
    return query


@reports_bp.get("/dashboard")
@login_required
def metrics_dashboard():
    record_type = request.args.get("record_type", "maintenance")
    if record_type == "fuel":
        query = select(FuelLoad).order_by(FuelLoad.loaded_at.desc())
        if request.args.get("technician_id", type=int): query = query.where(FuelLoad.technician_id == request.args.get("technician_id", type=int))
        if request.args.get("date_from"): query = query.where(FuelLoad.loaded_at >= request.args["date_from"])
        fuel_loads = db.session.execute(query).scalars().all()
        counts = {}
        for load in fuel_loads:
            user = db.session.get(User, load.technician_id); counts[user.full_name] = counts.get(user.full_name, 0) + 1
        technicians = db.session.execute(select(User).where(User.is_active == True).order_by(User.full_name)).scalars().all()  # noqa: E712
        return render_template("metrics_dashboard.html", record_type=record_type, fuel_loads=fuel_loads, reports=[], total=len(fuel_loads), by_user=sorted(counts.items(), key=lambda x:x[1], reverse=True), by_service={"Carga de petróleo":len(fuel_loads)}, by_machine={}, by_date={x.loaded_at.strftime("%d-%m"): sum(1 for y in fuel_loads if y.loaded_at.strftime("%d-%m")==x.loaded_at.strftime("%d-%m")) for x in fuel_loads}, technicians=technicians, areas=[], sections=[], machines=[])
    reports = db.session.execute(_filtered_reports().order_by(Report.created_at.desc())).scalars().all()
    counts = {}
    for report in reports:
        counts[report.technician.full_name] = counts.get(report.technician.full_name, 0) + 1
    service_counts, machine_counts, date_counts = {}, {}, {}
    for report in reports:
        service_counts[report.service_type] = service_counts.get(report.service_type, 0) + 1
        machine = report.machinery.name if report.machinery else "Sin maquinaria"
        machine_counts[machine] = machine_counts.get(machine, 0) + 1
        day = report.created_at.strftime("%d-%m") if report.created_at else "Sin fecha"
        date_counts[day] = date_counts.get(day, 0) + 1
    technicians = db.session.execute(select(User).where(User.is_active == True).order_by(User.full_name)).scalars().all()  # noqa: E712
    return render_template("metrics_dashboard.html", record_type=record_type, fuel_loads=[], reports=reports, total=len(reports), by_user=sorted(counts.items(), key=lambda x: x[1], reverse=True), by_service=service_counts, by_machine=sorted(machine_counts.items(), key=lambda x: x[1], reverse=True)[:8], by_date=date_counts, technicians=technicians, areas=db.session.execute(select(Area).order_by(Area.name)).scalars().all(), sections=db.session.execute(select(Section).order_by(Section.name)).scalars().all(), machines=db.session.execute(select(Machinery).order_by(Machinery.name)).scalars().all())


@reports_bp.get("/reports/new")
@login_required
def choose_report():
    return render_template("choose_report.html")

@reports_bp.route("/reports/maintenance/new", methods=["GET", "POST"])
@login_required
def create_report():
    if request.method == "POST":
        data = {
            "client": request.form.get("client", ""),
            "service_type": request.form.get("service_type", ""),
            "description": request.form.get("description", ""),
            "area_id": request.form.get("area_id"),
            "section_id": request.form.get("section_id"),
            "machinery_id": request.form.get("machinery_id"),
            "task_started_at": request.form.get("task_started_at", ""),
            "task_finished_at": request.form.get("task_finished_at", ""),
            "checklist": {key: request.form.get(f"check_{key}", "") for key, _, _ in CHECKLIST},
        }
        evidence_files = {key: request.files.get(f"evidence_{key}") for key, _, expected in CHECKLIST if expected}
        try:
            report = create_maintenance_report(
                data=data, technician_id=session["user_id"], evidence_files=evidence_files,
                attachments=request.files.getlist("attachments"), upload_folder=current_app.config["UPLOAD_FOLDER"],
            )
            flash("Reporte registrado y sellado con la hora del servidor.", "success")
            return redirect(url_for("reports.report_detail", report_id=report.id))
        except BusinessRuleError as error:
            flash(str(error), "error")
    return render_template("create_report.html", service_types=SERVICE_TYPES, checklist=CHECKLIST)

@reports_bp.route("/reports/fuel/new", methods=["GET", "POST"])
@login_required
def create_fuel_load():
    if request.method == "POST":
        data = {
            "loaded_at": request.form.get("loaded_at", ""), "observations": request.form.get("observations", ""),
            "generators": [{"number": number, "liters": request.form.get(f"liters_{number}"), "hourmeter": request.form.get(f"hourmeter_{number}")} for number in (1, 2)],
        }
        images = {(number, kind): request.files.get(f"{kind}_{number}") for number in (1, 2) for kind in ("water", "oil")}
        try:
            create_fuel_load_service(data=data, technician_id=session["user_id"], images=images, upload_folder=current_app.config["UPLOAD_FOLDER"])
            flash("Carga de petróleo registrada.", "success")
            return redirect(url_for("reports.dashboard"))
        except BusinessRuleError as error:
            flash(str(error), "error")
    return render_template("fuel_load.html")


@reports_bp.get("/api/areas")
@login_required
def areas_api():
    areas = db.session.execute(select(Area).where(Area.is_active == True).order_by(Area.name)).scalars().all()  # noqa: E712
    return jsonify([{"id": area.id, "name": area.name} for area in areas])


@reports_bp.get("/api/sections")
@login_required
def sections_api():
    area_id = request.args.get("area_id", type=int)
    if not area_id:
        return jsonify([])
    sections = db.session.execute(select(Section).where(Section.area_id == area_id, Section.is_active == True).order_by(Section.name)).scalars().all()  # noqa: E712
    return jsonify([{"id": section.id, "name": section.name} for section in sections])


@reports_bp.get("/api/machineries")
@login_required
def machineries_api():
    section_id = request.args.get("section_id", type=int)
    if not section_id:
        return jsonify([])
    machines = db.session.execute(select(Machinery).where(Machinery.section_id == section_id, Machinery.is_active == True).order_by(Machinery.name)).scalars().all()  # noqa: E712
    return jsonify([{"id": machine.id, "name": machine.name} for machine in machines])


@reports_bp.get("/reports/<int:report_id>")
@login_required
def report_detail(report_id):
    report = db.get_or_404(Report, report_id)
    checklist = db.session.execute(select(ReportChecklist).where(ReportChecklist.report_id == report.id).order_by(ReportChecklist.id)).scalars().all()
    return render_template("report_detail.html", report=report, checklist=checklist, image_extensions=IMAGE_EXTENSIONS)


@reports_bp.get("/reports/<int:report_id>/print")
@login_required
def print_report(report_id):
    report = db.get_or_404(Report, report_id)
    return render_template("report_print.html", report=report)


@reports_bp.get("/attachments/<int:attachment_id>")
@login_required
def download_attachment(attachment_id):
    attachment = db.get_or_404(Attachment, attachment_id)
    report = db.get_or_404(Report, attachment.report_id)
    if not _can_access_record(report.technician_id):
        return "Acceso no autorizado.", 403
    try:
        storage = get_file_storage()
        key = resolve_report_key(storage, attachment.report_id, attachment.stored_name)
        return _file_response(storage, key, attachment.original_name, attachment.mime_type, attachment=request.args.get("view") != "1")
    except StorageError:
        current_app.logger.exception("Error al recuperar adjunto %s", attachment.id)
        return "No fue posible recuperar el archivo.", 503


@reports_bp.get("/fuel-loads/<int:load_id>/generators/<int:generator_number>/images/<kind>")
@login_required
def view_fuel_image(load_id, generator_number, kind):
    if kind not in {"water", "oil"}:
        return "Archivo no encontrado.", 404
    load = db.get_or_404(FuelLoad, load_id)
    if not _can_access_record(load.technician_id):
        return "Acceso no autorizado.", 403
    detail = db.session.execute(select(FuelLoadGenerator).where(
        FuelLoadGenerator.load_id == load_id, FuelLoadGenerator.generator_number == generator_number
    )).scalar_one_or_none()
    if not detail:
        return "Archivo no encontrado.", 404
    stored_name = detail.water_image if kind == "water" else detail.oil_image
    try:
        storage = get_file_storage()
        key = resolve_fuel_key(storage, load_id, stored_name)
        return _file_response(storage, key, f"generador_{generator_number}_{kind}.png", "image/png", attachment=request.args.get("view") != "1")
    except StorageError:
        current_app.logger.exception("Error al recuperar fotografía de carga %s", load_id)
        return "No fue posible recuperar el archivo.", 503